Remove circular logic loop caused by programmatically setting current app without user click. (#226)
When the user clicks an app it was setting the local state selectedApp immediately and calling action setCurrentApp which will set the global current app after async all to bot completes.  However, the componentDidUpdate would fire after the local state change notice the selectedAPp did not match the global app and re-fire the call again with the old global current app.  This could cause never ending cycle depending on sequence of asynchronous calls returning before the local state changes.

*User now has to click on the app after creation.*

It might still be possible to automatically forward them to the app page, but for that we would need to await the call to create the new app and only once it's created do we make the calls to fetch all the data for thart particular app.  We could hack it into the apiHelper since the promises live there, but that is not good idea.