Update entity pickers to show all available tags instead of none when there is empty input better discovery. (#301)
Still don't think there is a way to show the dropdown when the click in the picker.  Only solution I could think of was to force a character into the input but I don't see a way to do that as this is not a controlled input. For now I made so that the entity trigger used in payload is ignored form search so typing '$' or a 'space' will show all options