Setup routing for app pages using NavLink instead of manual display actions (#319)
This uses the react router for switching app components like Entities, Actions, etc instead of manually controlling this with state. This allows full urls such as:
http://localhost:5050/home/c3c23b75-ac69-449b-89ec-10853d66e50c/actions

Ideally, when the page is refreshed we could reload the app based on the id in the url; however, this would require the user credentials but we don't have those. For now I put in a redirect back to the /home page to simulate behavior we have currently.

In future when we have full login I think we can do proper refresh and URL bookmarking.