Update app creation to default training status to Completed (#328)
Previously new apps had no training status given becuase it wasn't allowed to be sent to the server and it wasn't returned. This meant the TrainingStatus indicator defaulted to Unknown which could be misleading to the users. This will change it to default to Completed which is the expected neutral state.