test: adding train dialog test
Once the new model has been created, this test step validates the creation of a Train Dialog, and
then validates if the messages from the user and the bot are displayed in the chat pannel. Also,
this code change adds data-testids in preparation for new tests. Finally, the change also includes
the use of Components.