Tag and Frog and Fixes to Broken Tests (#856)
This is a start to the Tag and Frog scenario by creating the preliminary model necessary for the main test.
Also fixed a couple of bugs in test code:
The new time values in the grid broke tests that validated the dates that had been there.
There were two Refresh "Training Status" buttons on the page when Train Dialog modal is up. Cypress threw an error when trying to click on two buttons. Fixed it so it clicks on the last button.