fix: relax Poller test constraints to increase reliability
I think the issue might be related the differences in processing speed CI environments or variability in SetTimeout calls. Different environments can adjust them by a couple milliseconds so they're only approximate.
The odd thing is in this case it was *faster* instead of *slower* which you would not expect since they could throttle NodeJS timers or something.
The poller starts and has fixed interval of 100, we then add a poll immediately with expiration of 500.
The poller should invoked the methods at 100, 200, 300, 400
At 500 it should be expired and not call those methods thus leaving the count at 4, but somehow it was called 1 more time.
This could only happen if the last poll time was something like ~499