Failed Training Status, Loop Testing, Code Cleanup (#990)
This started out as changes to support Failed Training Status recovery so test can continue running.
I added support for running tests in a loop. I've wanted that for awhile now, and I needed it to test the failed recovery feature.
While I was at it I noticed multiple places in the code that could be cleaned up due to greater understanding of Cypress & JavaScript that I've gained since the code was written.