fix: use async await where possible (#998)
Not required but thought it is improvement to keep up with advances in typescript. Previously they didn't allow async/await when targeting es5 but 2.1 they do.
I think biggest benefit was standardized error handling. You can use same try/catch pattern for synchronous and asynchronous code

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-1.html

for the catch blocks I did `const error = e as Error` because of this:
https://github.com/Microsoft/TypeScript/issues/8677