Create End Session Model (#1103)
When I created the EndSession test case I manually created the model that I imported for that test thinking that I did not need to use automation to build it. However, I hit 2 significant bugs (1968 & 1969) at the time which reproduced for me in that short time frame I was working on it (over the course of 30 minutes or so). I filed the bugs and in the morning nobody, not even me, could reproduce the bugs.

So I am adding in this test case thinking if it is some confluence of timing issues with services that those bugs will eventually occur again and we can catch it.