fix: bug 2059 internal label confict errors (#1128)
* fix: bug 2059 internal label confict errors

Fix internal label conflict errors.  For some instances requires that the user save a partial edit dialog before continuing to edit. Eventually the server needs to not check for label conflicts for edited train dialogs when creating a teach session with context, but it currently does to that so we use a save pop-up as a workaround

* style: code review updates

* style: remove unneeded code