Fix Training Status Monitor (#1357)
Recently we have had multiple test consistently failing because the training status was not completing. It turns out it was in the queue for a long time, probably due to more tests, larger test models, and the recent change to run smoke tests in parallel with regression tests.

However, all of these theories might be wrong. And if they are then there is something else wrong with our servers and the better logging in this code will also help measure the issue and give us more information to help with fixing it.